app/requirements.txt
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
idna==2.6
Jinja2==2.10
jsonschema==2.6.0
oauthlib==2.0.6
pika==1.1.0
pybase64==1.0.2
PyJWT==1.5.3
pymongo==3.11.1
pytz==2020.4
requests==2.25.0
schedule==0.6.0
six==1.11.0
sqlparse==0.2.4
strgen==1.3.0
